+++ ros_sim/model_server/include/constant_model/ConstantModel.h
@@ -18,21 +18,21 @@ public:
 	/**
 	 * Initalize ConstantModel class with provided values
      */
-	ConstantModel(float u, float v, float temp, float salt, float dye, float depth);
+	ConstantModel(double u, double v, double temp, double salt, double dye, double depth);
 
 
 
-	const ModelData getData(float x, float y, float height, float time);
-	const ModelData getDataOutOfRange(float x, float y, float height, float time);
+	const ModelData getData(double x, double y, double height, double time);
+	const ModelData getDataOutOfRange(double x, double y, double height, double time);
 
 private:
 	
-	float u;
-	float v;
-	float temp;
-	float salt;
-	float dye;
-	float depth;
+	double u;
+	double v;
+	double temp;
+	double salt;
+	double dye;
+	double depth;
 
 };
